requirements.txt


pymongo==4.4.1
python-dotenv==1.0.0
requests==2.31.0
tweepy==4.14.0
xmltodict==0.13.0